requirements.txt
# requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==8.1.0
pandas==2.1.1
numpy==1.24.3
schedule==1.2.0
websockets==12.0